requirements.txt
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-api-key
inclusive-django-range-fields
ipython
drf-spectacular==0.10.0
pytz==2020.4
django-cors-headers
whitenoise==5.1.0
djoser==2.0.5
django_pandas
django-admin-rangefilter==0.6.3
django-admin-tools==0.9.0
requests
django-heroku==0.3.1
gunicorn==20.0.4
django-singleton-admin-2
django-better-admin-arrayfield
django-admin-numeric-filter
bpython # currently, Ipython has got a bug
silk
cookiecutter